Simple Mapper added.
The new mapper (simple mapper) does not use the Dot library
